added modification to PredatorPray and PredatorPrayTeleport scripts which ensures the cordinates cannot overflow

+++ PredatorPrayTeleport.java
@@ -13,16 +13,16 @@ public class PredatorPrayTeleport {
             System.exit(0);
         }
         // init and allocating memory. 
-        int[] pray_position = new int[2];         // accounts for overflow   
-        int[] predetor_position = new int[2];     // accounts for overflow
+        long[] pray_position = new long[2];         // long basic type accounts for overflow in the position of the position.   
+        long[] predetor_position = new long[2];     
 
         // declare start positions of pray
-        pray_position[0]=((int)Math.floor(Math.random()*((n-1)-0+1)));
-        pray_position[1]=((int)Math.floor(Math.random()*((n-1)-0+1)));
+        pray_position[0]=((long)Math.floor(Math.random()*((n-1)-0+1)));
+        pray_position[1]=((long)Math.floor(Math.random()*((n-1)-0+1)));
 
         // delcare start position of predetor
-        predetor_position[0]=((int)Math.floor(Math.random()*((n-1)-0+1)));
-        predetor_position[1]=((int)Math.floor(Math.random()*((n-1)-0+1)));
+        predetor_position[0]=((long)Math.floor(Math.random()*((n-1)-0+1)));
+        predetor_position[1]=((long)Math.floor(Math.random()*((n-1)-0+1)));
 
         // state parameters and starting positions
         System.out.format("n=%d s=%d t=%d",n,s,t);
@@ -40,11 +40,12 @@ public class PredatorPrayTeleport {
         }
     }
 
-    public static int[] new_position_pray(int step_size, int[] pray_position, int map_dimension){
+    public static long[] new_position_pray(int step_size, long[] pray_position, int map_dimension){
         int min = -step_size;
         int max = step_size;
-        pray_position[0]=((int)Math.floor(Math.random()*(max-min+1)+min))+pray_position[0];
-        pray_position[1]=((int)Math.floor(Math.random()*(max-min+1)+min))+pray_position[1];
+
+        pray_position[0]=((long)Math.floor(Math.random()*(max-min+1)+min))+pray_position[0];
+        pray_position[1]=((long)Math.floor(Math.random()*(max-min+1)+min))+pray_position[1];
 
         if(pray_position[0]>=map_dimension){
             pray_position[0]=pray_position[0]-(pray_position[0]-map_dimension+1);
@@ -66,7 +67,7 @@ public class PredatorPrayTeleport {
         return pray_position;
     }
     
-    public static int[] new_position_predetor(int step_size, int[] predetor_position, int[] pray_position, int map_size){
+    public static long[] new_position_predetor(int step_size, long[] predetor_position, long[] pray_position, int map_size){
         for (int i = 0; i < step_size; i++) {
             if(pray_position[0] > predetor_position[0]){
                 predetor_position[0]++;
@@ -87,10 +88,10 @@ public class PredatorPrayTeleport {
         System.out.format("[%d;%d]\n",predetor_position[0],predetor_position[1]);
         return predetor_position;
     }
-    public static int[] pray_teleport(int[] pray_position,int map_dimension){
+    public static long[] pray_teleport(long[] pray_position, int map_dimension){
         
-        pray_position[0]=((int)Math.floor(Math.random()*((map_dimension-1)-0+1)));
-        pray_position[1]=((int)Math.floor(Math.random()*((map_dimension-1)-0+1)));
+        pray_position[0]=((long)Math.floor(Math.random()*((map_dimension-1)-0+1)));
+        pray_position[1]=((long)Math.floor(Math.random()*((map_dimension-1)-0+1)));
 
         return pray_position;
     }
